+++ samples/NuvLocSample/AppShell.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace NuvLocSample;
+
+public partial class AppShell : Shell
+{
+	public AppShell()
+	{
+		InitializeComponent();
+	}
+}